requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dotenv==0.9.9
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
grpcio==1.76.0
humanfriendly==10.0
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lightning==2.5.6
lightning-utilities==0.15.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnx==1.16.1
onnx-ir==0.1.12
onnxruntime==1.23.2
onnxscript==0.1.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rich==14.2.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.5.1+cu121
torch-pruning==1.6.1
torchaudio==2.5.1+cu121
torchmetrics==1.8.2
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
Werkzeug==3.1.3
yarl==1.22.0
